requirements.txt
pydantic-ai>=0.0.14
pandas>=2.0
numpy>=1.24
python-dotenv>=1.0
openai>=1.0
fastapi>=0.109
uvicorn[standard]>=0.27
